ctf-task-book/requirements.txt
Flask
mysql-connector-python
Flask-Login
Werkzeug
python-dotenv